+++ magentoerpconnect/magento_model.py
@@ -52,12 +52,22 @@ class magento_backend(orm.Model):
 
     _backend_type = 'magento'
 
-    def _select_versions(self, cr, uid, context=None):
-        """ Available versions
+    def select_versions(self, cr, uid, context=None):
+        """ Available versions in the backend.
 
-        Can be inherited to add custom versions.
+        Can be inherited to add custom versions.  Using this method
+        to add a version from an ``_inherit`` does not constrain
+        to redefine the ``version`` field in the ``_inherit`` model.
         """
         return [('1.7', '1.7')]
+
+    def _select_versions(self, cr, uid, context=None):
+        """ Available versions in the backend.
+
+        If you want to add a version, do not override this
+        method, but ``select_version``.
+        """
+        return self.select_versions(cr, uid, context=context)
 
     def _get_stock_field_id(self, cr, uid, context=None):
         field_ids = self.pool.get('ir.model.fields').search(
@@ -407,7 +417,10 @@ class magento_store(orm.Model):
             'Create invoice on action',
             required=True,
             help="Should the invoice be created in Magento "
-                 "when it is validated or when it is paid in OpenERP?"),
+                 "when it is validated or when it is paid in OpenERP?\n"
+                  "This only takes effect if the sales order's related "
+                  "payment method is not giving an option for this by "
+                  "itself. (See Payment Methods)"),
     }
 
     _defaults = {