+++ magentoerpconnect/unit/backend_adapter.py
@@ -180,8 +180,19 @@ class GenericAdapter(MagentoCRUDAdapter):
 
         :rtype: dict
         """
+        arguments = [int(id)]
+        if attributes:
+            # Avoid to pass Null values in attributes. Workaround for
+            # https://bugs.launchpad.net/openerp-connector-magento/+bug/1210775
+            # When Magento is installed on PHP 5.4 and the compatibility patch
+            # (http://magento.com/blog/magento-news/magento-now-supports-php-54)
+            # is not installed, calling info() with None in attributes
+            # would return a wrong result (almost empty list of
+            # attributes). The right correction is to install the
+            # compatibility patch on Magento.
+            arguments.append(attributes)
         return self._call('%s.info' % self._magento_model,
-                          [int(id), attributes])
+                          arguments)
 
     def search_read(self, filters=None):
         """ Search records according to some criterias